+++ src/config/config.service.ts
@@ -42,6 +42,7 @@ class ConfigService {
 
   loadFromFile(configKey: string, filePath: string) {
     try {
+      // @ts-expect-error: Dynamically load the config file
       this.loadedConfig[configKey] = require(
         path.join(__dirname, filePath)
       ).default;
@@ -118,7 +119,7 @@ class ConfigService {
   }
 
   private loadCommands() {
-    const messageCommands = {};
+    const messageCommands: Record<string, Command> = {};
     const messageCommandsConfig = this.loadedConfig.messageCommands || {};
     Object.keys(messageCommandsConfig).forEach((key) => {
       messageCommands[key] = getCommandProperties(