+++ RELEASE_NOTES.md
@@ -1,5 +1,18 @@
 # Release Notes
 
+### 1.8.1 (2024-01-30)
+
+#### Features
+
+- New !wordle command to play wordle with the chat.
+
+#### Bug Fixes
+
+- Fixed Context.reply methods for twitch and discord.
+- Capped discord replies from openai as discord will fail on large messages.
+- Fixed `!poll` not sending the entire prompt.
+- More code cleanup.
+
 ### 1.8.0 (2024-01-29)
 
 #### Breaking Changes